fix: base build
