+++ .vscode/launch.json
@@ -1,0 +1,17 @@
+{
+    // Use IntelliSense to learn about possible attributes.
+    // Hover to view descriptions of existing attributes.
+    // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
+    "version": "0.2.0",
+    "configurations": [
+        {
+            "name": "debug build",
+            "request": "launch",
+            "type": "dart",
+            "flutterMode": "debug",
+            "args": [
+               
+            ]
+        }
+    ]
+}
