+++ lib/app.dart
@@ -1,0 +1,19 @@
+import 'package:flutter/material.dart';
+
+import 'pages/home/home_page.dart';
+
+class App extends StatelessWidget {
+  const App({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      title: 'Flutter Demo',
+      theme: ThemeData(
+        colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
+        useMaterial3: true,
+      ),
+      home: const HomePage(title: 'Flutter Demo Home Page'),
+    );
+  }
+}
